- partially added bcrypt support (still need a CSPRNG for every platform) - a couple other improvements

+++ src/crypto/blowfish.h
@@ -8,15 +8,15 @@ struct blowfish_ctx{
 
 void blowfish_init(struct blowfish_ctx *b);
 void blowfish_setkey(struct blowfish_ctx *b,
-			uint8 *key, uint32 len);
+			uint8 *key, size_t len);
 void blowfish_expandkey(struct blowfish_ctx *b,
-			uint8 *key, uint32 len);
+			uint8 *key, size_t len);
 void blowfish_expandkey1(struct blowfish_ctx *b,
-			uint8 *key, uint32 keylen,
-			uint8 *data, uint32 datalen);
-void blowfish_ecb_encode(struct blowfish_ctx *b, uint8 *data, uint32 len);
-void blowfish_ecb_decode(struct blowfish_ctx *b, uint8 *data, uint32 len);
-void blowfish_cbc_encode(struct blowfish_ctx *b, uint8 *iv, uint8 *data, uint32 len);
-void blowfish_cbc_decode(struct blowfish_ctx *b, uint8 *iv, uint8 *data, uint32 len);
+			uint8 *key, size_t keylen,
+			uint8 *data, size_t datalen);
+void blowfish_ecb_encode(struct blowfish_ctx *b, uint8 *data, size_t len);
+void blowfish_ecb_decode(struct blowfish_ctx *b, uint8 *data, size_t len);
+void blowfish_cbc_encode(struct blowfish_ctx *b, uint8 *iv, uint8 *data, size_t len);
+void blowfish_cbc_decode(struct blowfish_ctx *b, uint8 *iv, uint8 *data, size_t len);
 
 #endif //CRYPTO_BLOWFISH_H_
